SpringBoot2로 Rest api 만들기(7) –  MessageSource를 이용한 Exception 처리

+++ src/main/java/com/rest/api/repo/UserJpaRepo.java
@@ -3,5 +3,5 @@ package com.rest.api.repo;
 import com.rest.api.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserJpaRepo extends JpaRepository<User, Integer> {
+public interface UserJpaRepo extends JpaRepository<User, Long> {
 }
